Make installation instructions clearer
Thanks for making the type stubs for numpy!

It works but it took me a while to figure how to install this. This was the first stub I ever installed and it wasn't clear which folder I should put where. I cloned this whole repo into my `stubs` folder and it didn't work.

Maybe this little change will make it easier for others to install this stub.